feat: testing to prove that v2 works

+++ pkg/v2/rgb/rgb16bit_test.go
@@ -15,13 +15,13 @@ func TestRGB16Bit(t *testing.T) {
     }
 
     val, nextIdx := rgb.read(buffer, 0)
-    require.Equal(t, val, uint(0x45))
+    require.Equal(t, val, 0x45)
     require.Equal(t, nextIdx, 2)
     val, nextIdx = rgb.read(buffer, nextIdx)
-    require.Equal(t, val, uint(0x4545))
+    require.Equal(t, val, 0x4545)
     require.Equal(t, nextIdx, 4)
     val, nextIdx = rgb.read(buffer, nextIdx)
-    require.Equal(t, val, uint(0x4500))
+    require.Equal(t, val, 0x4500)
     require.Equal(t, nextIdx, 6)
 
 }
