+++ pkg/v2/rgb/rgb8bit_test.go
@@ -14,16 +14,16 @@ func TestRGB8Bit(t *testing.T) {
     }
 
     val, nextIdx := rgb.read(buffer, 0)
-    require.Equal(t, val, uint(0x0))
+    require.Equal(t, val, 0x0)
     require.Equal(t, nextIdx, 1)
     val, nextIdx = rgb.read(buffer, nextIdx)
-    require.Equal(t, val, uint(0x45))
+    require.Equal(t, val, 0x45)
     require.Equal(t, nextIdx, 2)
     val, nextIdx = rgb.read(buffer, nextIdx)
-    require.Equal(t, val, uint(0x45))
+    require.Equal(t, val, 0x45)
     require.Equal(t, nextIdx, 3)
     val, nextIdx = rgb.read(buffer, nextIdx)
-    require.Equal(t, val, uint(0x0))
+    require.Equal(t, val, 0x0)
     require.Equal(t, nextIdx, 4)
 }
